requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-treebeard==4.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.0.10
pandas==1.5.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
urllib3==1.26.13
uvicorn==0.20.0
whitenoise==6.2.0
